+++ lib/ui/view/UI5AntaresObjectPage.controller.ts
@@ -1,0 +1,67 @@
+import Button from "sap/m/Button";
+import MessageBox from "sap/m/MessageBox";
+import Controller from "sap/ui/core/mvc/Controller";
+import View from "sap/ui/core/mvc/View";
+import Targets from "sap/ui/core/routing/Targets";
+import { IObjectPageViewData } from "ui5/antares/types/entry/common";
+import { ODataMethods } from "ui5/antares/types/odata/enums";
+
+/**
+ * @namespace ui5.antares.ui.view
+ */
+export default class UI5AntaresObjectPage extends Controller {
+    private completeTriggered: boolean = false;
+
+    public onInit(): void {
+        const view = this.getView() as View;
+        const viewData = view.getViewData() as IObjectPageViewData;
+
+        view.addEventDelegate({
+            onAfterHide: () => {
+                if (!this.completeTriggered) {
+                    if (viewData.method === ODataMethods.DELETE) {
+                        const eventBus = viewData.entry.getSourceOwnerComponent().getEventBus();
+                        eventBus.publish("UI5AntaresEntryDelete", "UnsubscribeEvents");
+                    }
+
+                    viewData.entry.reset();
+                }
+                view.destroy();
+            }
+        });
+
+        if (view.byId("UI5AntaresObjectPageCompleteButton")) {
+            (view.byId("UI5AntaresObjectPageCompleteButton") as Button).attachPress({}, this.onComplete, this);
+        }
+
+        (view.byId("UI5AntaresObjectPageCancelButton") as Button).attachPress({}, this.onCancel, this);
+    }
+
+    public onComplete(): void {
+        this.completeTriggered = true;
+        const view = this.getView() as View;
+        const viewData = view.getViewData() as IObjectPageViewData;
+
+        if (viewData.method === ODataMethods.DELETE) {
+            const eventBus = viewData.entry.getSourceOwnerComponent().getEventBus();
+            eventBus.publish("UI5AntaresEntryDelete", "Complete");
+            eventBus.publish("UI5AntaresEntryDelete", "UnsubscribeEvents");
+        } else {
+            const validation = viewData.entry.valueValidation();
+
+            if (!validation.validated) {
+                MessageBox.error(validation.message);
+                return;
+            }
+
+            viewData.entry.submit();
+            (viewData.router.getTargets() as Targets).display(viewData.entry.getFromTarget());
+        }
+    }
+
+    public onCancel(): void {
+        const view = this.getView() as View;
+        const viewData = view.getViewData() as IObjectPageViewData;
+        (viewData.router.getTargets() as Targets).display(viewData.entry.getFromTarget());
+    }
+}