+++ CHANGELOG.md
@@ -4,6 +4,42 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
+## [1.96.25002] - 2024-06-07
+
+### Added
+
+- **Form Grouping Feature:** It is now possible to have form groups in the generated form. Additionally, form groups are used as the sections of the sap.uxap.ObjectPageLayout when the object page is activated.
+- **Object Page Feature:** It is now possible to decide to use either sap.m.Dialog or sap.uxap.ObjectPageLayout for Entry classes. 
+- **Manual Submit Feature:** It is now possible to submit the changes manually. The default behavior is still auto-submit by the library.
+
+### Changed
+
+- The auto-generated sap.m.Dialog now is draggable and resizable.
+
+## [1.120.1002] - 2024-06-07
+
+### Added
+
+- **Form Grouping Feature:** It is now possible to have form groups in the generated form. Additionally, form groups are used as the sections of the sap.uxap.ObjectPageLayout when the object page is activated.
+- **Object Page Feature:** It is now possible to decide to use either sap.m.Dialog or sap.uxap.ObjectPageLayout for Entry classes. 
+- **Manual Submit Feature:** It is now possible to submit the changes manually. The default behavior is still auto-submit by the library.
+
+### Changed
+
+- The auto-generated sap.m.Dialog now is draggable and resizable.
+
+## [1.123.1002] - 2024-06-07
+
+### Added
+
+- **Form Grouping Feature:** It is now possible to have form groups in the generated form. Additionally, form groups are used as the sections of the sap.uxap.ObjectPageLayout when the object page is activated.
+- **Object Page Feature:** It is now possible to decide to use either sap.m.Dialog or sap.uxap.ObjectPageLayout for Entry classes. 
+- **Manual Submit Feature:** It is now possible to submit the changes manually. The default behavior is still auto-submit by the library.
+
+### Changed
+
+- The auto-generated sap.m.Dialog now is draggable and resizable.
+
 ## [1.120.1001] - 2024-05-30
 
 ### Added
@@ -30,6 +66,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 - Initial productive release.
 
+[1.96.25002]: https://github.com/hasanciftci26/ui5-antares/compare/1.96.25001...1.96.25002?diff=split
+[1.120.1002]: https://github.com/hasanciftci26/ui5-antares/compare/1.120.1001...1.120.1002?diff=split
+[1.123.1002]: https://github.com/hasanciftci26/ui5-antares/compare/1.123.1001...1.123.1002?diff=split
 [1.120.1001]: https://github.com/hasanciftci26/ui5-antares/tree/1.120.1001
 [1.96.25001]: https://github.com/hasanciftci26/ui5-antares/tree/1.96.25001
 [1.123.1001]: https://github.com/hasanciftci26/ui5-antares/tree/1.123.1001